+++ packages/components/src/Menu/MenuItem.tsx
@@ -24,14 +24,20 @@
 
  */
 
-import { CompatibleHTMLProps } from '@looker/design-tokens'
+import {
+  CompatibleHTMLProps,
+  size,
+  space,
+  SizeProps,
+  SpaceProps,
+} from '@looker/design-tokens'
 import { IconNames } from '@looker/icons'
 import styled from 'styled-components'
 import React, { FC, ReactNode, useContext, useState, useEffect } from 'react'
 import { useID } from '../utils/useID'
 import { Icon } from '../Icon'
 import { MenuContext, MenuItemContext } from './MenuContext'
-import { MenuItemLayout, MenuItemLayoutGrid } from './MenuItemLayout'
+import { MenuItemLayout } from './MenuItemLayout'
 
 export interface MenuItemProps extends CompatibleHTMLProps<HTMLElement> {
   iconArtwork?: ReactNode
@@ -85,6 +91,7 @@ const MenuItemInternal: FC<MenuItemProps> = (props) => {
   }
 
   const { setOpen } = useContext(MenuContext)
+
   const handleOnClick = (event: React.MouseEvent<HTMLLIElement>) => {
     setFocusVisible(false)
     onClick && onClick(event)
@@ -115,7 +122,11 @@ const MenuItemInternal: FC<MenuItemProps> = (props) => {
       />
     ) : (
       renderIconPlaceholder && (
-        <div data-testid={`menu-item-${renderedIconID}-icon-placeholder`} />
+        <IconPlaceholder
+          size={24 / (compact ? 1.25 : 1)}
+          mr="xsmall"
+          data-testid={`menu-item-${renderedIconID}-icon-placeholder`}
+        />
       )
     )
 
@@ -127,19 +138,16 @@ const MenuItemInternal: FC<MenuItemProps> = (props) => {
       compact={compact}
       disabled={disabled}
       focusVisible={isFocusVisible}
-      hasIcon={Boolean(renderedIcon)}
       onBlur={handleOnBlur}
       onClick={handleOnClick}
       onKeyUp={handleOnKeyUp}
       className={className}
     >
       <Component href={href} role="menuitem" target={target}>
-        <MenuItemLayoutGrid>
-          {renderedIcon}
-          {children}
-        </MenuItemLayoutGrid>
+        {renderedIcon}
+        <span>{children}</span>
+        {detail && <Detail>{detail}</Detail>}
       </Component>
-      {detail && <Detail>{detail}</Detail>}
     </MenuItemLayout>
   )
 }
@@ -151,4 +159,11 @@ const Detail = styled.div`
   margin-left: auto;
   margin-right: ${({ theme: { space } }) => space.medium};
   padding-left: ${({ theme: { space } }) => space.large};
+`
+
+interface IconPlaceholderProps extends SizeProps, SpaceProps {}
+
+const IconPlaceholder = styled.div<IconPlaceholderProps>`
+  ${size}
+  ${space}
 `