+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.8] - 2020-08-02
+
+### Fixed
+
+- `DialogContent` patch for IE11 display bug
+- `MenuItem` patch for IE11 display bug
+
 ## [0.9.7] - 2020-07-27
 
 ### Added